combine thrasher step, remove custom validation
