+++ libs/newsletters-data-client/src/lib/schemas/newsletter-data-type.ts
@@ -43,14 +43,9 @@ export const thrasherOptionsSchema = z.object({
 		.array(
 			z
 				.object({
-					// TODO - these should be drop-downs populated from existing launched newsletters
-					// plus the draft currently being created
-					// TODO - this has specifically been defined as a triple-thrasher, rather than a
-					// multi-thrasher.  The vast majority of multi-thrashers are triple-thrashers, so
-					// this is suitable for the mvp, but occasionally more than 3 newsletters are required
-					thrasher1: z.string().optional().describe('left-hand thrasher'),
-					thrasher2: z.string().optional().describe('middle thrasher'),
-					thrasher3: z.string().optional().describe('right-hand-thrasher'),
+					thrasher1: z.string().optional().describe('newsletter name 1'),
+					thrasher2: z.string().optional().describe('newsletter name 2'),
+					thrasher3: z.string().optional().describe('newsletter name 3'),
 				})
 				.describe('multi-thrasher configuration'),
 		)
